refactor(nika-cli,nika-dap): the caps round three — arg surfaces to the verbs, inventory to nika-dap
The union with #662's evidence shell re-breached both caps: nika-cli at
15,111 prod LOC (cap 15,000 · merge-headroom target ≤14,900) and
main.rs at 1,536 lines (file cap 1,500). The descent law, third time:
compute descends, render stays, every call site reads unchanged and the
shipped UX is byte-identical (nika evidence · nika trace … · verify ·
anchor).

main.rs sheds 370 lines (1,536 → 1,166):

- The whole `trace` clap tree — TraceAction (12 variants incl. Anchor)
  + TraceArgs — descends to the new verbs::trace::action (the
  verbs::key::KeyAction precedent: the bin composes, the verbs own
  their arg types). Re-exported at verbs::trace::{TraceAction,TraceArgs}.
- The routing helpers: flow_verb and load_events (+ the tolerant-recover
  wrapper and its tests) to verbs::trace::manage; check_dispatch +
  CheckFlags to verbs::check::dispatch (lazy.rs and registry_args.rs
  re-point); the MCP surface (McpAction · McpTransportArg · mcp_verb) to
  verbs::mcp_pins; the Evidence arm back to a helper (fn-length law);
  the anchor arm to anchor_verb (same law — main and trace_verb were at
  106/100).
- A rebase-born DUPLICATION deleted: main.rs's own resolve_trace /
  announce_latest alongside manage's (one voice kept), and a
  second TraceArgs that a swap re-landed dead.

The crate budget sheds 334 prod LOC (15,111 → 14,777): the workspace
INVENTORY — the bounded walk (collect_workflow_paths · SKIP_DIRS), the
per-file check-ladder audit, and the cross-file rollups — descends to
the new nika_dap::inventory (the drift family's sibling: drift audits
one file's declarations against its body, this audits the workspace's
file SET). context.rs keeps the probe fold, the assembly, and both
renders; probe.rs re-exports the walk at the old public path
(pub fn → pub use, the only baseline line that changes shape).

Numbers (scripts/ci/check-crate-size.sh, prod LOC):
- nika-cli: 15,111 → 14,777 (merge headroom restored)
- nika-dap: 6,907 → 7,138 (far under its own cap)
- main.rs: 1,536 → 1,166 lines

public-api.txt: nika-cli gains the action/dispatch/mcp/manage surfaces
(probe::collect_workflow_paths becomes a re-export, the single
deletion); nika-dap appends the inventory block with the tracing
Instrument/WithSubscriber pair per pub type (the committed baseline's
canonical order). Both rendered with the CI-pinned 0.51.0 --omit
auto-trait-impls + the documented skew repairs (PolicyExt Sized+ ·
zvariant NoneValue restored for the three Default types that carry it).

Gates: cargo test --workspace --lib (305+137 green) · clippy
--all-targets -D warnings (0) · fmt --check · hygiene 36 green / 0 red
· cargo deny check ok · machete clean · crate-size green · fn-length
green · all 8 ci-ratchets green.
