+++ crates/nika-cli/src/verbs/context.rs
@@ -12,95 +12,21 @@
 //! usernames into agent transcripts) · every array CAPPED and says so
 //! (silent truncation reads as « covered everything »).
 
-use std::collections::BTreeMap;
 use std::fmt::Write as _;
-use std::path::Path;
 
-use nika_event::fold::{self, RunFact};
-use serde::Serialize;
+use nika_dap::inventory::{Rollups, Workspace, collect_workflows, rollups};
+use nika_event::fold::{self};
 
 use crate::display::theme::{Role, Theme};
 use crate::verbs::{VerbOutput, probe};
 
-/// Most workflow files one context call audits — beyond this an agent
-/// should scope by directory (the cap is REPORTED, never silent).
-const MAX_WORKFLOWS: usize = 100;
-
 /// Most recent run journals folded (newest first).
 const MAX_RUNS: usize = 20;
-
-/// Directory-walk budget (entries visited) — a context call must stay
-/// instant on a monorepo and never wander into dependency trees.
-const WALK_BUDGET: usize = 20_000;
-
-/// One workflow file's audited facts — verdicts, never contents.
-// Independent verdict FLAGS on a wire struct, not a state machine —
-// the same flags-are-flags exemption the Theme struct carries.
-#[allow(clippy::struct_excessive_bools)]
-#[derive(Debug, Clone, Serialize)]
-struct WorkflowFact {
-    /// Root-relative path (absolute paths leak usernames).
-    path: String,
-    /// The `workflow:` id when the file parses (`None` = parse failed).
-    workflow: Option<String>,
-    /// The full `is_clean` verdict (ten finding surfaces).
-    clean: bool,
-    /// Clean AND zero `native-first` hints (the `--native-strict` bar).
-    strict_clean: bool,
-    /// Total findings across the report's surfaces (0 when clean).
-    findings: usize,
-    tasks: usize,
-    waves: usize,
-    /// Task count per verb (`infer` · `exec` · `invoke` · `agent`) —
-    /// the shape an agent reasons about before reading the file.
-    verbs: BTreeMap<&'static str, usize>,
-    /// Worst-case bounded cost in USD (the ceiling — or the FLOOR part
-    /// when `cost_is_floor`).
-    cost_bounded_usd: f64,
-    /// `true` = at least one task is unbounded: the number above is a
-    /// FLOOR, not a ceiling (unknown stays unknown · never $0).
-    cost_is_floor: bool,
-    /// A `permits:` boundary is declared (default-deny active).
-    permits_declared: bool,
-}
-
-/// The workspace half of the aggregate (pure over collected facts).
-#[derive(Debug, Serialize)]
-struct Workspace {
-    root: String,
-    git: bool,
-    workflows: Vec<WorkflowFact>,
-    /// `true` when `MAX_WORKFLOWS` cut the inventory — scope by directory.
-    workflows_capped: bool,
-    /// How many the walk FOUND (≥ the emitted list when capped — the
-    /// K8s limit/continue rule: a cap without a total reads as « all »).
-    workflows_total_found: usize,
-    runs: Vec<RunFact>,
-    runs_capped: bool,
-    runs_total_found: usize,
-}
-
-/// Cross-file rollups an agent reasons about before touching anything.
-#[derive(Debug, Serialize)]
-struct Rollups {
-    workflows_total: usize,
-    workflows_clean: usize,
-    workflows_with_findings: usize,
-    /// Σ bounded worst-case USD across clean-parsing files.
-    cost_bounded_usd: f64,
-    /// Any file whose ceiling is a floor makes the SUM a floor too.
-    cost_is_floor: bool,
-    permits_declared: usize,
-    /// Σ recorded spend across the folded runs (only events that carried
-    /// a cost — unpriced calls are COUNTED, never priced at $0).
-    runs_cost_usd: f64,
-    runs_unpriced_calls: u64,
-}
 
 /// The `nika welcome --deep` arm. `json` emits `context_version: 1`.
 #[must_use]
 pub fn run(json: bool, theme: Theme) -> VerbOutput {
-    let root = Path::new(".");
+    let root = std::path::Path::new(".");
     let (facts, capped, found) = collect_workflows(root);
     let (runs, runs_capped, runs_found) =
         fold::fold_traces(&root.join(".nika").join("traces"), MAX_RUNS);
@@ -124,99 +50,6 @@ pub fn run(json: bool, theme: Theme) -> VerbOutput {
         return VerbOutput::ok(render_json(&workspace, &rollups, &probe));
     }
     VerbOutput::ok(render_human(&workspace, &rollups, &probe, theme))
-}
-
-/// Walk the root (bounded · dot/dep dirs skipped) and audit every
-/// `*.nika.yaml` through the in-process check ladder.
-fn collect_workflows(root: &Path) -> (Vec<WorkflowFact>, bool, usize) {
-    let mut paths = Vec::new();
-    let mut budget = WALK_BUDGET;
-    probe::collect_workflow_paths(root, root, 6, &mut budget, &mut paths);
-    paths.sort();
-    let found = paths.len();
-    let capped = found > MAX_WORKFLOWS;
-    paths.truncate(MAX_WORKFLOWS);
-    (
-        paths.iter().map(|p| audit(root, p)).collect(),
-        capped,
-        found,
-    )
-}
-
-/// Audit one file — parse + the check ladder, folded to verdict facts.
-/// A file that does not read or parse is an inventory row with
-/// `clean: false` (an agent must SEE it, not silently miss it).
-fn audit(root: &Path, rel: &Path) -> WorkflowFact {
-    let path = rel.display().to_string();
-    let unread = |findings: usize| WorkflowFact {
-        path: path.clone(),
-        workflow: None,
-        clean: false,
-        strict_clean: false,
-        findings,
-        tasks: 0,
-        waves: 0,
-        verbs: BTreeMap::default(),
-        cost_bounded_usd: 0.0,
-        cost_is_floor: false,
-        permits_declared: false,
-    };
-    let Ok(yaml) = std::fs::read_to_string(root.join(rel)) else {
-        return unread(1);
-    };
-    let Ok(wf) = nika_schema::parse(
-        &yaml,
-        nika_schema::FileId::new(0),
-        nika_schema::ParseMode::Strict,
-    ) else {
-        return unread(1);
-    };
-    let report = nika_schema::check(&wf);
-    let mut verbs: BTreeMap<&'static str, usize> = BTreeMap::default();
-    for task in &wf.tasks {
-        *verbs.entry(task.value.action.verb()).or_insert(0) += 1;
-    }
-    let findings = report.conformance.len()
-        + report.secret_leaks.len()
-        + report.secret_egresses.len()
-        + report.capability_escapes.len()
-        + report.schema_findings.len()
-        + report.schema_lints.len()
-        + report.unknown_tools.len()
-        + report.unknown_args.len()
-        + report.missing_args.len()
-        + report.gate_findings.len();
-    let native_first_hints = report
-        .hints
-        .iter()
-        .filter(|h| h.kind == "native-first")
-        .count();
-    WorkflowFact {
-        path,
-        workflow: wf.workflow.as_ref().map(|w| w.value.clone()),
-        clean: report.is_clean(),
-        strict_clean: report.is_clean() && native_first_hints == 0,
-        findings,
-        tasks: wf.tasks.len(),
-        waves: report.waves.len(),
-        verbs,
-        cost_bounded_usd: report.cost.bounded_total_usd,
-        cost_is_floor: report.cost.has_unbounded,
-        permits_declared: wf.permits.is_some(),
-    }
-}
-
-fn rollups(facts: &[WorkflowFact], runs: &[RunFact]) -> Rollups {
-    Rollups {
-        workflows_total: facts.len(),
-        workflows_clean: facts.iter().filter(|f| f.clean).count(),
-        workflows_with_findings: facts.iter().filter(|f| !f.clean).count(),
-        cost_bounded_usd: facts.iter().map(|f| f.cost_bounded_usd).sum(),
-        cost_is_floor: facts.iter().any(|f| f.cost_is_floor),
-        permits_declared: facts.iter().filter(|f| f.permits_declared).count(),
-        runs_cost_usd: runs.iter().filter_map(|r| r.cost_usd).sum(),
-        runs_unpriced_calls: runs.iter().filter_map(|r| r.unpriced_calls).sum(),
-    }
 }
 
 /// The versioned machine aggregate (`context_version: 1` · additive-only).
@@ -314,6 +147,11 @@ mod tests {
     use std::path::PathBuf;
 
     use super::*;
+    use std::collections::BTreeMap;
+
+    use nika_dap::inventory::WorkflowFact;
+    use nika_event::fold::RunFact;
+
     use crate::verbs::exit;
 
     fn plain() -> Theme {
