+++ crates/nika-dap/src/lib.rs
@@ -28,6 +28,7 @@ pub mod chain;
 pub mod drift;
 pub mod evidence;
 pub mod flow;
+pub mod inventory;
 pub mod otel;
 pub mod recover;
 pub mod reproduce;
